fix: private Route 반영
